requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
marshmallow==3.20.1
cryptography==41.0.7
